+++ src/hanami/src/core/processing/logical_host.h
@@ -57,6 +57,7 @@ class LogicalHost
     virtual void syncWithHost(Hexagon* hexagon) = 0;
 
     Hanami::ItemBuffer<Block> blocks;
+    Hanami::ItemBuffer<SynapseSection> sections;
 
     void addWorkerTaskToQueue(const Hanami::WorkerTask task);
     const Hanami::WorkerTask getWorkerTaskFromQueue();